refactor: modify book detail api
